mcp_server/requirements.txt
fastmcp>=3.0.0
aiohttp>=3.9.0
pydantic>=2.0.0
Pillow>=10.0.0
pyyaml>=6.0.0
